Ajout de quelques règles métier

+++ Backend/api/db_access/DbCnx.py
@@ -1,7 +1,7 @@
 from snowflake.connector import connect, DictCursor
 from dotenv import dotenv_values
 from business.User import UserInDB
-import json
+from datetime import datetime
 
 config = {**dotenv_values(".env_API")}
 
@@ -45,27 +45,30 @@ class UserDao():
             ctx.close()
 
         return list
+    
 
     @staticmethod
-    def get_permissions():
+    def is_userID_in_Users(user_id):
         ctx = DbCnx.get_db_cnx()
         cs = ctx.cursor(DictCursor)
         try:
-            request =  "SELECT * FROM USER_PERMISSION"
+            request =  "SELECT * FROM USERS"
             cs.execute(request)
             list = cs.fetchall()
             list = [user['USER_ID'] for user in list]
-
+            value = user_id in list
         finally:
             cs.close()
             ctx.close()
 
-        return list
-    
+        return value
+
+
     @staticmethod
     def get_user_permissions(user_id):
         ctx = DbCnx.get_db_cnx()
         cs = ctx.cursor(DictCursor)
+        
         try:
             request =  f"SELECT * FROM USER_PERMISSION WHERE user_id = '{user_id}'"
             cs.execute(request)
@@ -76,7 +79,25 @@ class UserDao():
             cs.close()
             ctx.close()
         return list
+
+
+    @staticmethod
+    def is_in_user_permission_id(user_id, permission_id):
+        ctx = DbCnx.get_db_cnx()
+        cs = ctx.cursor(DictCursor)
+        try:
+            request =  f"SELECT * FROM USER_PERMISSION WHERE user_id = '{user_id}'"
+            cs.execute(request)
+            list = cs.fetchall()
+            list = [permission['PERMISSION_ID'] for permission in list]
+            value = permission_id in list
+        finally:
+            cs.close()
+            ctx.close()
+        return value
     
+
+
     @staticmethod
     def get_user(user_id: str):
         users = UserDao.get_users()     
@@ -87,25 +108,36 @@ class UserDao():
                 user_dict['permissions'] = UserDao.get_user_permissions(user_id)
                 return UserInDB(**user_dict)
     
+
     @staticmethod
     def add_user(user):
         ctx = DbCnx.get_db_cnx()
         cs = ctx.cursor(DictCursor)
+
+        create_date=  datetime.now().strftime("%Y-%m-%d")
+        last_upd_date = datetime.now().strftime("%Y-%m-%d")
+
         try:
             request =  f"INSERT INTO users (user_id, pwd_hash, firstname, lastname, user_email, position, create_date, last_upd_date, active) "
-            request += f"VALUES ('{user.user_id}', '{user.pwd_hash}', '{user.firstname}', '{user.lastname}', '{user.user_email}', '{user.position}', '{user.create_date}', '{user.last_upd_date}', '{user.active}')"
+            request += f"VALUES ('{user.user_id}', '{user.pwd_hash}', '{user.firstname}', '{user.lastname}', '{user.user_email}', '{user.position}', '{create_date}', '{last_upd_date}', '{user.active}')"
 
-            cs.execute(request)            
+            cs.execute(request)
+
         finally:
             cs.close()
             ctx.close()
 
         return {'Message ' : "User added"}
 
+
     @staticmethod
-    def add_user_permission(user_id, permission_id):
+    def add_user_permission(user):
         ctx = DbCnx.get_db_cnx()
         cs = ctx.cursor(DictCursor)
+
+        user_id = user.user_id
+        permission_id = user.permission_id
+
         try:
             request =  f"INSERT INTO user_permission (user_id, permission_id) "
             request += f"VALUES ('{user_id}', '{permission_id}')"
@@ -116,52 +148,60 @@ class UserDao():
             ctx.close()
 
         return {'Message' : f"Permission {permission_id} successfully given to user {user_id}"}
+    
 
     @staticmethod
     def edit_user(user):
         ctx = DbCnx.get_db_cnx()
         cs = ctx.cursor(DictCursor)
 
-        request =  f"UPDATE USERS SET "
-        request += f"pwd_hash = '{user.pwd_hash}', "
-        request += f"firstname = '{user.firstname}', "
-        request += f"lastname = '{user.lastname}', "
-        request += f"user_email = '{user.user_email}', "
-        request += f"position = '{user.position}', "
-        request += f"last_upd_date = CURRENT_DATE, "
-        request += f"active = '{user.active}' "
-        request += f"WHERE user_id = '{user.user_id}';"
-        
+        user_items = user.__dict__
+
+        request = "UPDATE Users SET"
+        for i in user_items :
+            if user_items[i] != None :
+                request = request + f" {i} = '{user_items[i]}' ,"
+        request = request[:-1]
+        request = request + f"WHERE user_id = '{user.user_id}';"
+
         try:
-            cs.execute(request)        
+            cs.execute(request)
+
         finally:
             cs.close()
             ctx.close()
         
-        return {'Message' : "User and User_permission updated"}
+        return {'Message' : "User updated"}
+
 
     @staticmethod
     def delete_user(user_id):
         ctx = DbCnx.get_db_cnx()
         cs = ctx.cursor(DictCursor)
+        
         try:
             request = f"DELETE FROM USERS WHERE user_id = '{user_id}'"
             cs.execute(request)
-        finally:
-            cs.close()
-            ctx.close()
-
-        return {'Message' : "User deleted"}
-
-    @staticmethod
-    def delete_user_permissions(user):
-        ctx = DbCnx.get_db_cnx()
-        cs = ctx.cursor(DictCursor)
-        try :
-            request = f"DELETE FROM USER_PERMISSION WHERE user_id = '{user.user_id}'"
+            request = f"DELETE FROM USER_PERMISSION WHERE user_id = '{user_id}'"
             cs.execute(request)
         finally:
             cs.close()
             ctx.close()
-        
-        return {'Message' : f"Permissions for user {user.user_id} successfully deleted"}
+
+        return {'Message' : f"User {user_id} and permissions successfully deleted"}
+    
+
+    @staticmethod
+    def delete_user_permission(user):
+        ctx = DbCnx.get_db_cnx()
+        cs = ctx.cursor(DictCursor)
+
+        try :
+            request = f"DELETE FROM USER_PERMISSION WHERE user_id = '{user.user_id}' AND permission_id ='{user.permission_id}'"
+            print(request)
+            cs.execute(request)
+        finally:
+            cs.close()
+            ctx.close()
+
+        return {'Message' : f"Permission {user.permission_id} for user {user.user_id} successfully deleted" }
